Update footer.html
Build was failing locally. These changes fixed the issue. The reason was that .File.Path is not always possible to resolve, especially when making the background stuff for the theme, hence it will fail. The implemented solution will make sure .File.Path can be resolved to something meaningful before doing so.